requirements.txt
argcomplete==1.12.3
boto==2.49.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudpickle==1.6.0
cox==0.1.post3
crcmod==1.7
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
dill==0.3.3
fasteners==0.16
gcs-oauth2-boto-plugin==2.7
gitdb==4.0.7
GitPython==3.1.17
google-apitools==0.5.32
google-reauth==0.1.1
GPUtil==1.4.0
grpcio==1.37.1
gsutil==4.62
gym==0.18.3
httplib2==0.19.1
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
learn2learn==0.1.5
matplotlib==3.4.2
mock==2.0.0
monotonic==1.6
networkx==2.5.1
numexpr==2.7.3
numpy==1.20.3
oauth2client==4.1.3
pandas==1.2.4
pbr==5.6.0
Pillow==8.2.0
protobuf==3.17.0
psutil==5.8.0
py3nvml==0.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.15
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-box==5.3.0
python-dateutil==2.8.1
pytz==2021.1
pyu2f==0.1.5
requests==2.25.1
retry-decorator==1.1.1
robustness==1.2.1.post2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
smmap==4.0.0
tables==3.6.1
tensorboardX==2.2
threadpoolctl==2.1.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
xmltodict==0.12.0